replace everything in optimizer dir with what was in optimizer_new
