--- optimizer/tools/__init__.py
@@ -1,6 +1,0 @@
-from .knowledge_base_rag_tool import (
-    knowledge_base_rag_tool,
-    create_knowledge_base_rag_tool,
-)
-
-__all__ = ["knowledge_base_rag_tool", "create_knowledge_base_rag_tool"]
